Add files via upload
Prime Factors 
If you arrangement the prime factors of the natural numbers and place them in columns (shown above) with the lowest vale on the left, and increasing in value as you move to the right, another pattern emerges that reveals the primes.
Notice, the spacing of the first occurrence of each prime is double that in the preceding column. eg. for 2 it goes 2,4,8,16,.... For 3 it is 3,6,12,24.... For 5 it is 5,10,20,40.... For 7 it is 7,14,28,56..... etc

Take the first column only
Notice that for each prime their occurrence begins with the prime number itself and then the second occurrence is the p-1, there after the patterns deviate somewhat and become more complex as the prime values increase. The spacing of 5 and 7 are somewhat  simpler to follow.  Using the prime 5 the pattern is easy 4,2,4,2……
Using the prime 7 it is the pattern 6,4,2,4,2,4,6,2 …. That is continually repeated. Unfortunately I did not have enough data to explore the patterns fully. However the limite